% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_attach_policy}
\alias{clouddirectory_attach_policy}
\title{Attaches a policy object to a regular object}
\usage{
clouddirectory_attach_policy(DirectoryArn, PolicyReference,
  ObjectReference)
}
\arguments{
\item{DirectoryArn}{[required] The Amazon Resource Name (ARN) that is associated with the Directory
where both objects reside. For more information, see arns.}

\item{PolicyReference}{[required] The reference that is associated with the policy object.}

\item{ObjectReference}{[required] The reference that identifies the object to which the policy will be
attached.}
}
\description{
Attaches a policy object to a regular object. An object can have a
limited number of attached policies.
}
\section{Request syntax}{
\preformatted{clouddirectory$attach_policy(
  DirectoryArn = "string",
  PolicyReference = list(
    Selector = "string"
  ),
  ObjectReference = list(
    Selector = "string"
  )
)
}
}

\keyword{internal}
